Stop holding the account key between actions
The signing session was created and never ended, so the derived key sat in page
memory until the tab closed. Deriving it costs a touch of a finger, so there was
never a reason to keep it.

Every entry point now zeroes its session in a finally block. register() and
restore() return an address and nothing else. withSigner() lends a signer for
one piece of work and ends it after, which makes a stray reference inert rather
than dangerous — Mera throws SESSION_ENDED on any later use. Session.open(ttl)
covers a multi-step flow and closes itself, exposing expiresAt so the interface
can warn before it does.

A per-transaction key would be the wrong shape: the key is the account, so a
fresh one each time means a fresh address each time and money stranded at the
previous identity. The claim-link key is the case where per-use is right, and
it already is.

+++ scripts/lifecycle.ts
@@ -131,7 +131,7 @@ async function main() {
 
   // ---- the recipient can find it with nothing stored ----------------------
   const incoming = await read(iris, I, "incomingOf", [recipient.address]);
-  check(incoming.length === 1 && incoming[0] === id, "commitment is discoverable from the address alone");
+  check(incoming.length === 1 && BigInt(incoming[0]) === id, "commitment is discoverable from the address alone");
 
   // ---- nothing is due yet -------------------------------------------------
   // Compare against chain time rather than assuming the preceding transactions
@@ -221,7 +221,7 @@ async function main() {
   await send(iris, I, "cancel", [id]);
   const after = await read(ausd, A, "balanceOf", [sponsor.address]);
   c = await read(iris, I, "get", [id]);
-  line("refunded to sender", usd(after - before));
+  line("refunded to sender", usd((after as bigint) - (before as bigint)));
   check(c.cancelled === true, "commitment is cancelled");
   check(
     (await read(ausd, A, "balanceOf", [iris])) === 0n,
